Move FormulaUtils.getValueByPath() to a Maps utility class
Since the getValueByPath() function could be useful to walk any map with
String keys, move it to a utility class.
